[examples] Rename inspect to inspector
This commit renames inspect module to inspector.

TICO-DCO-1.0-Signed-off-by: seongwoo <mhs4670go@naver.com>
